Global update.
Controllers :
Student : Add the function to display the theme(s) available depending on the classroom of the current user. ( Use a temporary fix to the issue #4)
Teacher: Include a temporary fix for the issue #4)

Entities :
Hello to the rescue entity "classHasTheme". (For mor info about why this entity is here check issue #4)

Template :
tl;dr