Bringing a structure update.
The actual structure was not good so we bring a small update.

With this new structure Themes is a group of Levels, Levels contain various quizz related to the current Theme.

Bringing a new progression system.
With the new structure, a student will need to answers to Level 1 to have access to Level 2 (seem legit). Levels can contain a various number of quiz but when student select a Level, a quiz is randomly selected.