Huge update.
Controllers :

On TeacherController lots of change and lots of new things:
themeSummaryAction() have been reworked.
addThemeAction() takes lots of hotfix but it's not enough for the moment :/ I'm still working on it to fix this ASAP...
addLevelsAction() & addQuizzAction() are two new method not coded yet. Will be done when addThemeAction () will be corrected.

Entities :

Looooooooooots changes in the Entities but there is more to come.

Template :

Basic stuff. No reel change in it.